% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{maps_layout_classic}
\alias{maps_layout_classic}
\title{Create a Classic Map Layout for Quebec}
\usage{
maps_layout_classic(map_all, map_mtl, map_qc)
}
\arguments{
\item{map_all}{A ggplot object representing the full Quebec map.}

\item{map_mtl}{A ggplot object representing the Montreal RMR map.}

\item{map_qc}{A ggplot object representing the Quebec City map.}
}
\value{
A ggplot object with the combined layout.
}
\description{
This function combines three map visualizations (Quebec, Montreal RMR, Quebec City)
into a classic layout. The Quebec map takes the larger left part, while the Montreal
RMR and Quebec City maps are placed side by side on the right.
}
\examples{
maps_layout_classic(map_all, map_mtl, map_qc)
}
