% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map_minimal}
\alias{theme_map_minimal}
\title{Apply a clean map theme to ggplot objects}
\usage{
theme_map_minimal(legend = FALSE, background_color = NULL)
}
\arguments{
\item{legend}{logical; if TRUE, includes the legend at the bottom}
}
\value{
A ggplot2 theme
}
\description{
This theme removes axes, ticks, and text, providing a clean map view.
It optionally includes a legend at the bottom if specified.
}
\examples{
ggplot(data) +
  geom_sf() +
  theme_map_minimal(legend = TRUE)
}
