% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_map.R
\name{zoom_map}
\alias{zoom_map}
\title{Zoom into a specified geographic area on a map}
\usage{
zoom_map(long_min, long_max, lat_min, lat_max)
}
\arguments{
\item{long_min}{Minimum longitude for the map view.}

\item{long_max}{Maximum longitude for the map view.}

\item{lat_min}{Minimum latitude for the map view.}

\item{lat_max}{Maximum latitude for the map view.}
}
\value{
A list containing coord_sf parameters for ggplot2.
}
\description{
This function adjusts the map view to focus on a specific geographic area
defined by minimum and maximum longitudes and latitudes.
}
\examples{
zoom_map(-74, -73.45, 45.3975, 45.70)
}
